Little change in the event interface stuff
Added a Call method to IEvent and made it void.
The thing with Call returning the event object is that it
needs to be cast at the user end anyway and nobody does that.
And this way Call is available in the interface.

Will come in useful very soon.
Just you wait!